Moved config for passport and mongo to new config file
